requirements.txt

streamlit
numpy
Pillow
scikit-learn
